refactor(system): expose get-{schema,navigation} as utilities
