+++ source/application/routes/meta.js
@@ -1,16 +1,9 @@
-import {
-	resolve
-} from 'path';
-import getPatternTree from '../../library/utilities/get-pattern-tree';
+import getNavigation from '../../library/get-navigation';
 
 export default (application, configuration) => {
 	return async function metaRoute() {
-		const config = application.configuration[configuration.options.key];
-		const {patterncwd, cwd} = application.runtime;
-		const path = resolve(patterncwd || cwd, config.path);
+		console.log(configuration.options);
 		this.type = 'json';
-		this.body = await getPatternTree('.', path, {
-			cache: application.cache
-		});
+		this.body = await getNavigation(application, configuration.options.key);
 	};
 };
